Added more structure and diagrams that describe lexer states and transitions
